+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ### (????-??-??) v0.22
 
+- Fix `JsonGenerator` state corruption in `ExceptionResolver` if
+  `LogEvent#getThrown()` is `null`.
+
 - Add `level:severity` and `level:severity:code` resolvers. (#48)
 
 - Add `timestamp:divisor=<divisor>` resolver. (#48)